requirements.txt
fastapi
pydantic
python-dotenv
spacy
uvicorn
redis
motor